+++ ChangeLog.md
@@ -14,6 +14,8 @@ XP Framework Core ChangeLog
 
 ### Features
 
+* Added `io.FileUtil::append()` to complement the `write()` method.
+  (@thekid)
 * Added optional initial bytes to `io.streams.OutputStream` constructor
   (@thekid)
 * Added `read()` and `write()` methods to `io.FileUtil` as replacements